requirements.txt
PyQt6>=6.4.0
mutagen>=1.46.0
python-vlc>=3.0.18121
spotipy>=2.23.0